backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pyjwt
python-multipart
